nbits.h

+++ src/algo/hash/_hash_.c
@@ -1,5 +1,6 @@
 #include "_hash_.h"
 #include <nbits.h>
+#include <nint.h>
 #include <stdio.h>
 
 
@@ -7,5 +8,5 @@ void
 print_hash(const char *name, const char *ss, unsigned long hash)
 {
   printf("%s\n------------\n%s -> %16lu\n" BPRI64 "\n", name, ss, hash,
-         BIT64(hash));
+         BIT64((uint64_t) hash));
 }
